requirements.txt
langchain
pydantic
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
python-dotenv
duckduckgo-search